Update to v00r06 release.
byuu says:

New-style S-SMP syntax is completed. I really, really like it. Hopefully
people will give it a fair chance, but I suppose it's as likely as
headerless ROMs catching on. Which is a shame, I think it makes reading
and writing SPC700 ASM a hell of a lot easier. The chip was very clearly
a ripoff of the 6502, and these mnemonics make it a lot more clear, and
easier for 6502 coders to use it.

I now have my 'modern' syntax for the CPU and SMP as snes-cpu and
snes-smp. New versions allow an opcode name suffix of < (byte),
> (word), ^ (long), which act as spaces. This allows you to have
vertically aligned code with no extraneous spacing. Old version keeps
the tried and true .b, .w, .l syntax. New version unifies 65816 and
SPC700 syntax to use + for indexing instead of commas, canonical version
sticks to what you're used to.

New option added to the assembler: -case-insensitive.
If you set this, it will auto-lowercase your entire file upon loading it
in. This has no effect on text inside 'apostrophes' or "quotes".