Update to v02r08 release.
byuu says:

Updated the documentation. Defines are now argumentless macros instead
of redefinable labels. The reason for doing that is because labels are
meant to be used before being defined, and thus cannot be redefined
later. By trying to shoehorn defines onto them for the nicer {}-less
syntax, it allows the same undeclared define to go beyond the first
pass. Anyway, as a bonus for having to use {}, the evaluation happens
post-expansion, allowing for the "lda #$00" -> byte-size detection to
work, whereas lda #offset's size cannot be determined, as offset has
already been evaluated to a raw integer.