Update to v04r01 release.
(Not sure what happened to v04. The announcement post was edited to
describe v04r01 soon after it was posted)

byuu says:

This release should fix an issue with the isbyte/isword/islong closures.
It also creates a new category that I'm calling intrinsics: built-in
macros with special meanings. The +/- relative label stuff is now bound
to it as well, for instance:

    mapper lorom
    org $8000; fill $8000

    org $8000; base $c000
      -; bra {-1}
      -; bra {-2}
      -; bra {-3}
      bra {+1}; +
      bra {+2}; +
      bra {+3}; +

      -; stz $55aa
      ?; stz $55aa
      ~; stz $55aa

      bra {-3}
      bra {-2}
      bra {-1}
      bra {-}

      bra {+}
      bra {+1}
      bra {+2}
      bra {+3}

      ~; stz $55aa
      ?; stz $55aa
      +; stz $55aa

      print "PC: {origin} [{pc}]"
      jmp {pc}

+ and - now both evaluate using the same relative label index counter,
meaning they are functionally identical.
I only kept + and - because it can make it clearer whether you intend
a label to be jumped back to or forward to.

I've also added ? and ~ as alternate ways to declare anonymous labels.
Consider them temporary, I just want to see what people think. At best,
I'll only keep one. At worst, I'll remove both.

-----

EDIT: okay, removed ?, I will keep ~, - and +. Converted {pc} to {$},
and {origin} to {@}. This means that all intrinsics are automatically
reserved and will never conflict with legal user defines/macros.
[I don't actually stop you from creating eg macro
!#@$@ざけんなよ、パーサー！！#^$%&(&*), but you really shouldn't do that
:P]