Update to v04r04 release.
(v04r03 was not posted to the bass release thread)

byuu says:

Added a new architecture, which I'm just calling "table" for lack of
a better name (suggestions welcome.)

Basically, it lets you load in "arch" files that declare instruction
sets. And in fact, you can load in multiple arch specs in one pass, so
CPU+SMP code in the same file will be possible once again.

I was using something close to table-driven, but I was handling the
parsing for each addressing mode by hand.
With this new version, each table entry takes a list that tells you how
to assemble the data.
You can actually write individual bits at a time, which is necessary for
some of the SMP and Cx4 instructions.
Basically, relationships that are too complex to represent in one entry
just need more than one.
For instance, adc #$0000 and adc #$00. Store one for each mode.

One of the nice things of the new parser is that you can have spaces
inside the evaluated part, eg "lda #3 * (5 + 2)"
Amusingly, the table parser + complete 65816 instruction set is <12KB of
code, whereas the plain 65816 module by itself was 20KB of code.

I still need to write the SMP table, so for now only the CPU table can
be tested.

The table assembler is just an arch, so it inherits from the Bass core
just like any other arch. This will allow us to still add specialty
archs for platforms where a table is just never going to be versatile
enough (hi, x86/amd64.)
The most important part of building a table is that the order of entries
is also the order of evaluation.
lda (*,x) must come before lda (*), which must come before lda *.
Otherwise, the latter will match the former pairs.