Update to v05r04 release.
(r03 was not posted to the bass thread)

byuu says:

At last, got it. You can now define a macro based on a condition, or
a condition inside of a macro.
If you get stupid and try nesting them too deeply, I don't know if it
will work or not. But the code is getting too complex for me to fix it
if it doesn't work. My mind can only keep track of so many variables at
one time.