Update to v10 release.
v10 was released on 2013-01-16. byuu says:

Came up with a macro expansion that works in O(n) time. Unfortunately it's not super fast. A ten-line macro recursion needs about one second per 10,000 iterations. That's honestly to be expected with vector growth and manual string parsing and evaluation with every pass, with no caching of tokens or anything. Was still hoping it'd be a bit better, but oh well. Works out to about 100,000 lines of code parsed per second. Compared to v09.02's evaluator though, it's far faster. 1 second vs 16 seconds for the 10K iterations, and 100K would basically never complete thanks to quadratic growth.