Update to v11r04 release.
v11r01 was released 2013-04-17. byuu said:

So this is an attempt to extend bass to be more than just a cross-assembler.
The idea being that we can use custom archs to perform DSL tasks. Not really sure what I have in mind just yet, though.
"-o output" is now an optional command-line argument. Attempting to write to a file that isn't opened will produce an error.
There is also the new "output "filename", mode" directive. "mode" is optional (it can be create or modify, defaults to the command-line setting: -overwrite ? create : modify)
So with this, you can use "bass file.asm", and have the ASM file specify which file to write to. Or if you want, which files.
I also added the "constant" directive, which creates labels. You can use them to create labels for functions inside a binary you are modifying, or you can use them for static, read-only computation. You cannot redefine them. They are to labels as defines are to macros, basically.
Code:
  lda addr,x  //lda $208000,x
  constant base $200000
  constant addr base + $8000
  if addr == $208000; ...; endif  //be careful, the first pass won't know the value of addr until it has been declared

v11r02 was released 2013-04-22. byuu said:

Implemented red-black trees with O(log n) insert/remove/search as nall::set.
Labels and Macros now use nall::set instead of nall::vector.
The internal release name for bass v12 is hereby "striped bass" :)

v11r04 was released 2013-05-05. byuu said:

This release should get bass as fast as it's ever going to get.
nall::string now supports empty strings, size+capacity (can seek to end of string in O(1) time), and copy-on-write semantics.
Macros have also been overhauled one last time. We are back to only one type of macro evaluation, yet the recursion remains. I've added the "return" statement that can be specified within macros to specify what a macro should be substituted with in its parent block invocation. Macro evaluation also now starts with the deepest macro first, which allows macros to accept macros as arguments once again.
The end result is, in my opinion, quite elegant. At least compared to what was necessary before.
Code:
  macro add x,y
  return {eval {self::x} + {self::y}}
  endmacro

  macro factorial n
  if {self::n} == 1
    return 1
  else
    return {eval {self::n} * {factorial {eval {self::n} - 1}}}
  endif
  endmacro

  macro echo text
  print {self::text}
  endmacro

  main:
  print "{add {add 1,2}, {add 3,4}}! = {factorial 10}"
  {echo "hello, world!"}
I'm still not a huge fan of "{self::*}", but I'm hard-pressed to think of any alternatives. The only symbols I have left that can be used as a shortcut for "self::" are ` and @, both of which are hideous. Note that whatever we use has to work for both macros and labels, so that's why we can't just use eg $. That would collide with eval() trying to parse it as a hex value.
eval is also a necessary evil. Not all return results can be evaluated, such as the "echo" example. Attempting to auto-evaluate a string would end badly. Plus there may be times when you want to return "number+number" instead of "eval(number+number)"
Here I could reduce it slightly, eg:
Code:
  return {={self::x}+{self::y}}
  vs
  return {eval {self::x}+{self::y}}
Not sure whether terseness or expressiveness is better in this case, and I'm not going to offer both.
This would be the terse alternative:
Code:
  macro add x,y
  return {={@x} + {@y}}
  endmacro

  macro factorial n
  if {@n} == 1
    return 1
  else
    return {={@n} * {factorial {={@n} - 1}}}
  endif
  endmacro

  macro echo text
  print {@text}
  endmacro

  main:
  print "{add {add 1,2}, {add 3,4}}! = {factorial 10}"
  {echo "hello, world!"}