Update to v11r06 release.
v11r06 was released on 2013-05-07. byuu said:

Parsing should be even faster now, thanks to small-string-optimization.
Macro arguments are now referenced as {#argument}, and local macros and labels can be created with the # prefix as well.
Escape sequences have been added, so you can express more text:
Code:
  define '\n' 0xff
  db "Hero of time,\n", "face me!\0"
With that, I pretty much had to move print to not auto-append a line-feed character.
Fixed skeud's reported issue (thanks for the heads up.)
The last thing I'd like to look at before finalizing on bass v12 is macro arguments.
It would be really nice to let a macro argument be a macro that gets evaluated later, and then provide the option to evaluate it immediately with {eval ...}, but pulling that off is easier said that done. It's not the way evaluation works at all, so it'd have to be special cased and inconsistent with the rest of the system.