Update to v14r06 release.
v14r06 was released on 2017-06-29. byuu said:

Changelog:
* `scope {` removed -- didn't appear to serve any purpose
* `scope name {` is now `namespace name {`
* `scope name: {` is now `function name {`
* fixed `global` prefix ... it wasn't working at all before
* arch snes.cpu: added tcs/tsc variants for tas/tsa
* macros are now always scoped no matter what

The thing with macros and scoping is ... they create their own stack frames for variables and such. Once the macro execution concludes, the frames are destroyed. Even if that weren't the case, macros support arity overriding, so you can have `test()` and `test(arg1, arg2)` and internally they become `test:0` and `test:2`. If you want a macro to create a variable outside of its execution stack frame, then you'll want to prefix the variable with `global` to do that.

I updated the documentation with this WIP as well to reflect the changes.