Update to v14r08 release.
v14r08 was released on 2017-08-16. byuu said:

On the spirit of breaking changes, we may as well get them all in now. arch was the only abbreviated keyword in bass, and specifying root systems is kind of silly and wouldn't work with eg the Z80 which could appear in the Master System, Game Gear and Mega Drive.

So this is now:

  arch snes.cpu => architecture wdc65816
  arch snes.smp => architecture spc700
  arch reset => architecture reset

This will allow us to grow and add more architectures more easily going forward.

I'm also debating whether or not to enforce quotes around architecture names. This could be a way to have a unified syntax between built-in architectures and external ones. Or maybe it's best to just externalize them instead of bundling them into bass in the first place. This could be important as others create their own architecture tables, which don't ship with bass official. But if in a fork they're built-in, then the syntax becomes impossible to use with the official bass.

Next up, I've rearranged the command-line arguments. You now need:

  bass (create|modify) [options] source [source ...]

Basically, -create is no longer a flag with the default option being to modify existing files.

Now it's explicit and fixed as the first argument.

My reasoning is: there's really not a superior option as a default, with the other option being uncommon. Both are equally likely. I pretty much never use modify mode now that I'm writing my own programs for the 21fx instead of ROM hacking existing games. In cases of equal likelihood, I believe it's better to be specific about what you want. And in this way, it's also future proofed. Say we wanted to change the default to create, and have -modify, that'd be very bad. By having both, we sidestep the issue. No hidden knowledge of which option is the favored one is necessary.

> So macros would have symbols scoped by default, and inlines would be unscoped?

Correct.

> But can there be a way to do what global did in r06? Maybe uh...well, can't think of any good names for it.

It's possible. I'm also unsure whether inline should affect both framing and scoping.

  constant PPU_CTRL1        ($2000)
  constant PPU_CTRL2        ($2001)
  constant PPU_STATUS       ($2002)
  constant PPU_SPR_ADDR     ($2003)
  constant PPU_SPR_IO       ($2004)
  constant PPU_VRAM_ADDR1   ($2005)
  constant PPU_VRAM_ADDR2   ($2006)
  constant PPU_VRAM_IO      ($2007)

I went ahead and supported this as well. Whitespace can now appear between name(parameters) now in everything that takes (parameters).