Update to v15 release.
v15 was released on 2017-10-08. byuu says:

After two years, a new version of bass has been released, which contains many substantial changes: expressions, multi-argument defines, inline macros, and much more.

Unfortunately, code written for bass v14 is unlikely to be compatible with v15, and code changes will be necessary.

However, the goal going forward is for bass to stabilize. I wanted to offer a guarantee that all code compiled with v15 will be forward-compatible, but given the large number of changes, this seems a bad idea. Instead, I'd like to run with v15 for a decent length of time, and if no serious issues arise, then we can consider this for the next release.

Do note that unlike previous releases, backward-incompatible changes won't occur to v15 unless they are of significant, quantifiable benefit.